feat(converter): implement ASCII and binary STL format conversion logic
Add support for converting 3D mesh files between ASCII STL and Binary STL formats in Go.
This update introduces core parsing, tokenization/lexing, and binary encoding routines
to process STL files efficiently with minimal memory overhead.

Key Changes:
- Implement STL parser for ASCII (`solid ... endsolid`) and Binary headers/facets.
- Add support for normal vectors and vertex Coordinate (X, Y, Z) extraction.
- Introduce reader and writer interfaces to facilitate streaming conversion.
- Add error handling for malformed STL files and unexpected EOF encountered during parsing.
- Add unit test coverage for ASCII-to-Binary and Binary-to-ASCII conversion routines.

+++ internal/converter/binary_defs.go
@@ -25,7 +25,9 @@ type BinaryFacet struct {
 	Attrib Attribute
 }
 
-func CalculateFileSize(numTriangles uint32) uint { return HEADER_SIZE + TRIANGLE_COUNT_SIZE + (numTriangles * FACET_SIZE) }
+func CalculateFileSize(numTriangles uint32) uint {
+	return uint(uint64(HEADER_SIZE) + uint64(TRIANGLE_COUNT_SIZE) + uint64(numTriangles)*uint64(FACET_SIZE))
+}
 
 // --- //
